parse clusters by id works
