+++ parser/parse_by_api.py
@@ -2,48 +2,37 @@
 ## Using School 21 API to get peers data.
 """
 
-import os
+import logging
 import requests
+import get_auth_token
 
 
-def get_token():
+def parse_clusters():
     """
-    ###  Auth and return token: `Bearer $token`
+    ### Getting info about online peers on campuses maps
     """
-    request = requests.post(
-        url="https://auth.sberclass.ru/auth/realms/EduPowerKeycloak/protocol/openid-connect/token",
-        data={
-            "username": os.getenv("EDU_SCHOOL_LOGIN"),
-            "password": os.getenv("EDU_SCHOOL_PASSWORD"),
-            "grant_type": "password",
-            "client_id": "s21-open-api",
-        },
-    )
+    kzn_clusters = {
+        "34734": "et",
+        "34735": "ev",
+        "34736": "ge",
+        "34737": "pr",
+        "34738": "si",
+        "34739": "un",
+        "34740": "va",
+    }
 
-    # requests.post()
-    print(request)
-    # print(request.json())
+    token = get_auth_token.get_token()
+    for cluster_id, cluster_name in kzn_clusters.items():
+        request = requests.get(
+            url=f"https://edu-api.21-school.ru/services/21-school/api/v1/clusters/{cluster_id}/map?limit=101&offset=0&occupied=true",
+            headers={"accept": "application/json", "Authorization": token},
+            timeout=5,
+        )
 
-    token = request.json()["access_token"]
-
-    token = "Bearer " + token
-    return token
-
-
-def parse():
-    """
-    ### Getting info about peers on campuses maps
-    """
-    token = get_token()
-
-    request = requests.get(
-        url="https://edu-api.21-school.ru/services/21-school/api/v1/campuses",
-        headers={"accept": "application/json", "Authorization": token},
-    )
-    print(request)
-    print(request.json())
+        print(request)
+        print(request.json())
 
 
 if __name__ == "__main__":
 
-    parse()
+    parse_clusters()
